Update package app-eselect/eselect-opengl -Added support for cleaning up GL lib symlinks -Now uses modified locations for both GL libs & Xorg drivers Updated customised/patched media-libs/mesa package Added customised/patched x11-base/xorg-server package Added customised/patched x11-drivers/nvidia-drivers package Added profiles package.mask file: to mask these 4 packages Updated profiles categories file

+++ app-eselect/eselect-opengl/eselect-opengl-1.3.2.ebuild
@@ -15,9 +15,7 @@ HOMEPAGE="https://www.gentoo.org/"
 GLEXT="85"
 GLXEXT="34"
 
-MIRROR="https://dev.gentoo.org/~mattst88/distfiles"
-SRC_URI="https://dev.gentoo.org/~mgorny/dist/opengl.eselect-${PV}.xz"
-#	${MIRROR}/${P}.tar.xz"
+SRC_URI=""
 
 LICENSE="GPL-2"
 SLOT="0"
@@ -26,25 +24,18 @@ IUSE=""
 
 DEPEND="app-arch/xz-utils"
 RDEPEND=">=app-admin/eselect-1.2.4
-		 !<media-libs/mesa-11.0.0
+		 >=media-libs/mesa-11.0.6-r1
+		 >=x11-base/xorg-server-1.16.4-r6
 		 !<x11-proto/glproto-1.4.17-r1
-		 !<=x11-base/xorg-server-1.12.4-r5
-		 !=x11-base/xorg-server-1.15.2-r1
-		 !=x11-base/xorg-server-1.15.2-r2
-		 !~x11-base/xorg-server-1.16.3
-		 !=x11-base/xorg-server-1.16.4
-		 !=x11-base/xorg-server-1.16.4-r3
-		 !~x11-base/xorg-server-1.17.0
-		 !=x11-base/xorg-server-1.17.1
 		 !<x11-drivers/ati-drivers-14.9-r2
 		 !=x11-drivers/ati-drivers-14.12
 		 !<=app-emulation/emul-linux-x86-opengl-20140508"
 
-S=${WORKDIR}
+S="${WORKDIR}"
 
 pkg_preinst() {
 	# we may be moving the config file, so get it early
-	OLD_IMPL=$(eselect opengl show)
+	OLD_IMPLEMENTATION=$(eselect opengl show)
 }
 
 pkg_postinst() {
@@ -55,13 +46,11 @@ pkg_postinst() {
 		find "${EROOT}"/usr/lib*/opengl -depth -type d -empty -exec rmdir -v {} +
 	fi
 
-	if [[ -n "${OLD_IMPL}" && "${OLD_IMPL}" != '(none)' ]] ; then
-		eselect opengl set "${OLD_IMPL}"
+	if [[ -n "${OLD_IMPLEMENTATION}" && "${OLD_IMPLEMENTATION}" != '(none)' ]] ; then
+		eselect opengl set "${OLD_IMPLEMENTATION}"
 	fi
-	if [[ -f ${EROOT}/etc/env.d/03opengl ]]; then
-		# remove the old file, moved now
-		rm -vf "${EROOT}"/etc/env.d/03opengl
-	fi
+	[[ -f "${EROOT}"/etc/env.d/03opengl ]] && rm -vf "${EROOT}"/etc/env.d/03opengl
+	[[ -f "${EROOT}"/etc/X11/xorg.conf.d/20opengl.conf ]] && rm -vf "${EROOT}"/etc/X11/xorg.conf.d/20opengl.conf
 }
 
 src_prepare() {
@@ -69,18 +58,23 @@ src_prepare() {
 	if [[ ${CHOST} == *-darwin* ]] ; then
 		sed -i -e 's/libGL\.so/libGL.dylib/' opengl.eselect || die
 	fi
-	epatch "${FILESDIR}/${P}-fix-gl-provider-lib-support.patch"
 }
 
 src_install() {
 	insinto "/usr/share/eselect/modules"
-	newins opengl.eselect-${PV} opengl.eselect
-#	doman opengl.eselect.5
+	newins "${FILESDIR}/opengl.eselect-${PV}" opengl.eselect
 }
 
 pkg_postinst() {
-		ewarn "This is an experimental version of ${CATEGORY}/${PN} designed to fix various issues"
-		ewarn "when switching GL providers."
-		ewarn "This package can only be used in conjuction with a specially patched version"
-		ewarn "of media-libs/mesa ."
+	ewarn "This is an experimental version of ${CATEGORY}/${PN} designed to fix various issues"
+	ewarn "when switching GL providers."
+	ewarn "This package can only be used in conjuction with patched versions of:"
+	ewarn " * media-libs/mesa"
+	ewarn " * app-select/eselect-opengl"
+	ewarn " * x11-base/xorg-server"
+	ewarn "from the bobwya overlay."
+	ewarn ""
+	ewarn "Please ensure you cleanup the /usr/lib{32,64}/ lib*GL* symlinks, using:"
+	ewarn "  eselect opengl cleanup"
+	ewarn "... before reverting back to the mainline Gentoo implementation of this package"
 }
